+++ src/GameDetect/Detection/IFullscreenDetector.cs
@@ -1,0 +1,7 @@
+namespace GameDetect.Detection;
+
+public interface IFullscreenDetector
+{
+    bool IsFullscreen();
+}
+