+++ src/GameDetect/Detection/IProcessMonitor.cs
@@ -1,0 +1,9 @@
+using GameDetect.Detection.Models;
+
+namespace GameDetect.Detection;
+
+public interface IProcessMonitor
+{
+    List<ProcessInfo> GetGameCandidates();
+}
+