+++ src/GameModeAPI/Detection/IFullscreenDetector.cs
@@ -1,6 +1,0 @@
-namespace GameModeAPI.Detection;
-
-public interface IFullscreenDetector
-{
-    bool IsFullscreen();
-}