+++ src/GameModeAPI/Detection/IProcessMonitor.cs
@@ -1,8 +1,0 @@
-using GameModeAPI.Detection.Models;
-
-namespace GameModeAPI.Detection;
-
-public interface IProcessMonitor
-{
-    List<ProcessInfo> GetGameCandidates();
-}