+++ src/GameModeAPI/State/IStateManager.cs
@@ -1,8 +1,0 @@
-using GameModeAPI.Detection.Models;
-
-namespace GameModeAPI.State;
-
-public interface IStateManager
-{
-    (GameState State, bool Changed) UpdateState(DetectedGame? detected);
-}