+++ tests/GameDetect.Tests/UnitTest1.cs
@@ -1,0 +1,10 @@
+namespace GameDetect.Tests;
+
+public class UnitTest1
+{
+    [Fact]
+    public void Test1()
+    {
+
+    }
+}